delete obsolete creative templates and related code
