--- static/src/javascripts/projects/commercial/views/creatives/fabric-expandable-video-v2-cta.html
@@ -1,1 +1,0 @@
-<a class="creative__cta creative__cta--position-<%=position%> <%=media%>" href="<%=link%>" target="_blank"><img src="<%=image%>" alt></a>
