--- static/src/javascripts/projects/commercial/views/creatives/fabric-expandable-video-v2.html
@@ -1,54 +1,0 @@
-<div id="<%= data.id %>" class="creative--expandable-video creative--expandable-video-v2 <%=data.videoOptions%>">
-
-    <div class="ad-slot__label adFullWidth__label facia-container--layout-front">
-        <div class="facia-container__inner">Advertisement</div>
-    </div>
-
-    <div class="ad-exp--expand l-side-margins mobile-only" style="background: <%=data.backgroundColor%> url('<%=data.backgroundImageM%>') <%=data.backgroundRepeatM%> <%=data.backgroundPositionM%>">
-        <div class="slide-container">
-            <div class="ad-exp-collapse__slide slide-1" style="background: <%=data.slide1BGColor%> url('<%=data.slide1BGImageM%>') <%=data.slide1BGImageRepeatM%> <%=data.slide1BGImagePositionM%>;">
-                <div class="ad-exp__open">
-                    <div class="ad-exp__layer ad-exp__layer1" style="background: url('<%=data.slide1Layer1BGImageM%>') <%=data.slide1Layer1BGImageRepeatM%> <%=data.slide1Layer1BGImagePositionM%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer2" style="background: url('<%=data.slide1Layer2BGImageM%>') <%=data.slide1Layer2BGImageRepeatM%> <%=data.slide1Layer2BGImagePositionM%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer3" style="background: url('<%=data.slide1Layer3BGImageM%>') <%=data.slide1Layer3BGImageRepeatM%> <%=data.slide1Layer3BGImagePositionM%>;"></div>
-                </div>
-            </div>
-            <div class="ad-exp-collapse__slide slide-2" style="background: <%=data.slide2BGColor%> url('<%=data.slide2BGImageM%>') <%=data.slide2BGImagePositionM%> <%=data.slide2BGImageRepeatM%>;">
-                <div class="ad-exp__layer ad-exp__layer1" style="background: url('<%=data.slide2Layer1BGImageM%>') <%=data.slide2Layer1BGImageRepeatM%> <%=data.slide2Layer1BGImagePositionM%>;"></div>
-                <div class="ad-exp__layer ad-exp__layer2" style="background: url('<%=data.slide2Layer2BGImageM%>') <%=data.slide2Layer2BGImageRepeatM%> <%=data.slide2Layer2BGImagePositionM%>;"></div>
-                <div class="ad-exp__layer ad-exp__layer3" style="background: url('<%=data.slide2Layer3BGImageM%>') <%=data.slide2Layer3BGImageRepeatM%> <%=data.slide2Layer3BGImagePositionM%>;"></div>
-            </div>
-        </div>
-    </div>
-
-    <div class="ad-exp--expand l-side-margins hide-until-tablet" style="background: <%=data.backgroundColor%> url('<%=data.backgroundImage%>') <%=data.backgroundRepeat%> <%=data.backgroundPosition%>">
-        <div class="slide-container" style="">
-            <div class="ad-exp-collapse__slide slide-1" style="background: <%=data.slide1BGColor%> url('<%=data.slide1BGImage%>') <%=data.slide1BGImageRepeat%> <%=data.slide1BGImagePosition%>;">
-                <div class="ad-exp__open">
-                    <div class="ad-exp__layer ad-exp__layer1" style="background: url('<%=data.slide1Layer1BGImage%>') <%=data.slide1Layer1BGImageRepeat%> <%=data.slide1Layer1BGImagePosition%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer2" style="background: url('<%=data.slide1Layer2BGImage%>') <%=data.slide1Layer2BGImageRepeat%> <%=data.slide1Layer2BGImagePosition%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer3" style="background : url('<%=data.slide1Layer3BGImage%>') <%=data.slide1Layer3BGImageRepeat%> <%=data.slide1Layer3BGImagePosition%>;"></div>
-                </div>
-            </div>
-            <div class="ad-exp-collapse__slide slide-2" style="background: <%=data.slide2BGColor%> url('<%=data.slide2BGImage%>') <%=data.slide2BGImagePosition%> <%=data.slide2BGImageRepeat%>;">
-                <div class="ad-exp__layer ad-exp__layer1" style="background: url('<%=data.slide2Layer1BGImage%>') <%=data.slide2Layer1BGImageRepeat%> <%=data.slide2Layer1BGImagePosition%>;"></div>
-                <div class="ad-exp__layer ad-exp__layer2" style="background: url('<%=data.slide2Layer2BGImage%>') <%=data.slide2Layer2BGImageRepeat%> <%=data.slide2Layer2BGImagePosition%>;"></div>
-                <div class="ad-exp__layer ad-exp__layer3" style="background: url('<%=data.slide2Layer3BGImage%>') <%=data.slide2Layer3BGImageRepeat%> <%=data.slide2Layer3BGImagePosition%>;"></div>
-            </div>
-        </div>
-    </div>
-
-    <div class="slide-video">
-        <div class="slide-container">
-            <a href="<%=data.clickMacro%><%=data.link%>" target="_blank" class="video-container__cta"></a>
-            <div class="video-container__inner">
-                <%=data.videoEmbed%>
-            </div>
-            <%=data.mobileCTA%>
-            <%=data.desktopCTA%>
-        </div>
-    </div>
-
-    <%=data.showPlus%>
-    <%=data.showArrow%>
-</div>
