--- static/src/javascripts/projects/commercial/views/creatives/fabric-expanding-v1.html
@@ -1,48 +1,0 @@
-<div id="<%= data.id %>" class="creative--expandable creative--fabric-expanding-v1">
-    <div class="ad-slot__label adFullWidth__label facia-container--layout-front">
-        <div class="facia-container__inner">Advertisement</div>
-    </div>
-    <div class="ad-exp--expand l-side-margins mobile-only" style="background-color: <%=data.backgroundColor%>">
-        <div class="facia-container__inner facia-container__inner--full-span" style="background: url('<%=data.backgroundImageM%>') <%=data.backgroundRepeatM%> <%=data.backgroundPositionM%>;">
-            <%=data.showPlus%>
-            <%=data.showArrow%>
-            <div class="ad-exp-collapse__slide slide-1" style="background: <%=data.slide1BGColor%> url('<%=data.slide1BGImageM%>') <%=data.slide1BGImageRepeatM%> <%=data.slide1BGImagePositionM%>;">
-                <a href="<%=data.clickMacro%><%=data.link%>" target="_new">
-                    <div class="ad-exp__layer ad-exp__layer1" style="background: url('<%=data.slide1Layer1BGImageM%>') <%=data.slide1Layer1BGImageRepeatM%> <%=data.slide1Layer1BGImagePositionM%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer2" style="background: url('<%=data.slide1Layer2BGImageM%>') <%=data.slide1Layer2BGImageRepeatM%> <%=data.slide1Layer2BGImagePositionM%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer3" style="background: url('<%=data.slide1Layer3BGImageM%>') <%=data.slide1Layer3BGImageRepeatM%> <%=data.slide1Layer3BGImagePositionM%>;"></div>
-                </a>
-            </div>
-            <div class="ad-exp-collapse__slide slide-2" style="background: <%=data.slide2BGColor%> url('<%=data.slide2BGImageM%>') <%=data.slide2BGImagePositionM%> <%=data.slide2BGImageRepeatM%>;">
-                <%=data.videoMobile%>
-                <a href="<%=data.clickMacro%><%=data.link%>" target="_new">
-                    <div class="ad-exp__layer ad-exp__layer1" style="background: url('<%=data.slide2Layer1BGImageM%>') <%=data.slide2Layer1BGImageRepeatM%> <%=data.slide2Layer1BGImagePositionM%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer2" style="background: url('<%=data.slide2Layer2BGImageM%>') <%=data.slide2Layer2BGImageRepeatM%> <%=data.slide2Layer2BGImagePositionM%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer3" style="background: url('<%=data.slide2Layer3BGImageM%>') <%=data.slide2Layer3BGImageRepeatM%> <%=data.slide2Layer3BGImagePositionM%>;"></div>
-                </a>
-            </div>
-        </div>
-    </div>
-    <div class="ad-exp--expand l-side-margins hide-until-tablet" style="background-color: <%=data.backgroundColor%>">
-        <%=data.scrollbg%>
-        <div class="facia-container__inner facia-container__inner--full-span" style="background: url('<%=data.backgroundImage%>') <%=data.backgroundRepeat%> <%=data.backgroundPosition%>;">
-            <%=data.showPlus%>
-            <%=data.showArrow%>
-            <div class="ad-exp-collapse__slide slide-1" style="background: <%=data.slide1BGColor%> url('<%=data.slide1BGImage%>') <%=data.slide1BGImageRepeat%> <%=data.slide1BGImagePosition%>;">
-                <a href="<%=data.clickMacro%><%=data.link%>" target="_new">
-                    <div class="ad-exp__layer ad-exp__layer1" style="background: url('<%=data.slide1Layer1BGImage%>') <%=data.slide1Layer1BGImageRepeat%> <%=data.slide1Layer1BGImagePosition%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer2" style="background: url('<%=data.slide1Layer2BGImage%>') <%=data.slide1Layer2BGImageRepeat%> <%=data.slide1Layer2BGImagePosition%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer3" style="background : url('<%=data.slide1Layer3BGImage%>') <%=data.slide1Layer3BGImageRepeat%> <%=data.slide1Layer3BGImagePosition%>;"></div>
-                </a>
-            </div>
-            <div class="ad-exp-collapse__slide slide-2" style="background: <%=data.slide2BGColor%> url('<%=data.slide2BGImage%>') <%=data.slide2BGImagePosition%> <%=data.slide2BGImageRepeat%>;">
-                <%=data.videoDesktop%>
-                <a href="<%=data.clickMacro%><%=data.link%>" target="_new">
-                    <div class="ad-exp__layer ad-exp__layer1" style="background: url('<%=data.slide2Layer1BGImage%>') <%=data.slide2Layer1BGImageRepeat%> <%=data.slide2Layer1BGImagePosition%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer2" style="background: url('<%=data.slide2Layer2BGImage%>') <%=data.slide2Layer2BGImageRepeat%> <%=data.slide2Layer2BGImagePosition%>;"></div>
-                    <div class="ad-exp__layer ad-exp__layer3" style="background: url('<%=data.slide2Layer3BGImage%>') <%=data.slide2Layer3BGImageRepeat%> <%=data.slide2Layer3BGImagePosition%>;"></div>
-                </a>
-            </div>
-        </div>
-    </div>
-</div>
