--- static/src/javascripts/projects/commercial/views/creatives/fabric-expanding-video.html
@@ -1,8 +1,0 @@
-<iframe
-    width="<%=width%>"
-    height="<%=height%>"
-    src="<%=src%>"
-    frameborder="0"
-    class="<%=className%>"
-    style="<%=inlineStyle%>">
-</iframe>
