--- static/src/javascripts/projects/commercial/views/creatives/fabric-v1-video.html
@@ -1,9 +1,0 @@
-<iframe
-    width="409px"
-    height="230px"
-    src="<%=videoURL%>?rel=0&amp;controls=0&amp;showinfo=0&amp;title=0&amp;byline=0&amp;portrait=0"
-    frameborder="0"
-    class="fabric-v1_video fabric-v1_video--desktop fabric-v1_video--vert-pos-<%=videoPositionV%> fabric-v1_video--horiz-pos-<%=videoPositionH%>"
-    style="<%=position%>">
-</iframe>
-
