--- static/src/javascripts/projects/commercial/views/creatives/scrollbg.html
@@ -1,1 +1,0 @@
-<div class="ad-scrolling-bg" style="background-image: url('<%=backgroundImageP%>'); background-position: 50% 0; background-repeat: <%=backgroundImagePRepeat%>"></div>
